@@ -16,7 +16,8 @@
   "dependencies": {
     "@slack/bolt": "^3.8.1",
     "@slack/web-api": "^6.5.0",
-    "firebase": "^9.4.1",
+    "axios": "^0.24.0",
+    "firebase": "^9.8.4",
     "firebase-admin": "^9.8.0",
     "firebase-functions": "^3.14.1"
   },
@@ -27,7 +28,14 @@
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-import": "^2.22.0",
     "firebase-functions-test": "^0.2.0",
-    "typescript": "^3.8.0"
+    "typescript": "^4.5.4"
   },
-  "private": true
+  "private": true,
+  "version": "1.0.0",
+  "directories": {
+    "lib": "lib"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": ""
 }
