@@ -8,29 +8,42 @@
     "start": "next start"
   },
   "dependencies": {
+    "@cloudinary/react": "^1.0.0-beta.14",
+    "@cloudinary/url-gen": "^1.3.0",
+    "@reduxjs/toolkit": "^1.8.1",
     "@slack/bolt": "^3.8.1",
     "@slack/web-api": "^6.5.0",
+    "@types/react-scroll": "^1.8.3",
     "browser-image-compression": "^1.0.17",
+    "caniuse-lite": "^1.0.30001323",
     "clsx": "^1.1.1",
+    "dompurify": "^2.3.4",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
     "firebase-admin": "^10.0.0",
     "firebase-functions": "^3.16.0",
-    "next": "^12.0.4",
+    "next": "^12.1.5",
     "prettier": "^2.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.19.5",
+    "react-hook-form": "^7.33.1",
+    "react-redux": "^8.0.1",
+    "react-scroll": "^1.8.6",
     "react-select": "^5.2.1",
     "react-slick": "^0.28.1",
     "react-tsparticles": "^1.31.2",
+    "react-twitter-embed": "^3.0.3",
     "react-youtube": "^7.13.1",
+    "sanitize-html": "^2.6.1",
     "slick-carousel": "^1.8.1",
     "swr": "^0.5.6",
     "vercel": "^23.1.2"
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
+    "@swc/cli": "^0.1.57",
+    "@swc/core": "^1.2.174",
+    "@types/dompurify": "^2.3.2",
     "@types/node": "^15.0.2",
     "@types/react": "^17.0.5",
     "@types/react-dom": "^17.0.3",
