@@ -1,6 +1,8 @@
-import {getAuth} from "firebase/auth";
-import {initializeApp, getApps} from "firebase/app";
+import { getAuth } from "firebase/auth";
+import { initializeApp, getApps } from "firebase/app";
 import { config } from "../index";
+import { getStorage } from "firebase/storage";
+import { getFirestore } from "firebase/firestore";
 
 const firebaseConfig = {
   apiKey: config.fb_project_config.firebase_api_key,
@@ -12,10 +14,11 @@ const firebaseConfig = {
   measurementId: config.fb_project_config.firebase_mesuarement_id,
   databaseUrl: config.fb_project_config.firebase_database_url,
 };
-console.log(process.env.FIREBASE_API_KEY);
 
 const apps = getApps();
 
 const firebaseApp = !apps.length ? initializeApp(firebaseConfig) : apps[0];
 
 export const auth = getAuth(firebaseApp);
+export const storage = getStorage(firebaseApp)
+export const db = getFirestore(firebaseApp)
