@@ -1,25 +1,26 @@
 import * as functions from "firebase-functions";
-import { SignUpData } from "./type";
+import { FormDataType, SignUpAuthorizationDataTypeDataWithImageId } from "./type";
 
 export const config = functions.config()
 
 import { expressReceiver } from './slack/app'
-import usePostMessage from './slack/postMessage'
+import postFormData from './slack/postFormData'
+import postNewUserData from "./slack/postNewUserData";
 
 export const slack1 = functions.https.onRequest(expressReceiver.app)
 
-export const postMessageToSlackChannelWithUserData1 = functions.https.onCall(
-  (data: SignUpData, context) => {
+export const postNewUserDataToSlack = functions.https.onCall(
+  (data: SignUpAuthorizationDataTypeDataWithImageId, context) => {
     console.log('context', context)
     console.log('data on firebase', data)
-    return usePostMessage(data)
+    return postNewUserData(data)
   }
 )
 
-// // Start writing Firebase Functions
-// // https://firebase.google.com/docs/functions/typescript
-//
-// export const helloWorld = functions.https.onRequest((request, response) => {
-//   functions.logger.info("Hello logs!", {structuredData: true});
-//   response.send("Hello from Firebase!");
-// });
+export const postFormDataToSlack = functions.https.onCall(
+  (formData: FormDataType, context) => {
+    console.log('context', context)
+    console.log('formdata on firebase')
+    return postFormData(formData)
+  }
+)
