@@ -1,22 +1,33 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-import {SignUpData} from "../type";
 
 //  ユーザー確認メッセージを表示するBlockを作成
-export const createMessageBlock = (data: SignUpData) => {
-  const {name, email, password} = data;
+import { SignUpAuthorizationDataTypeDataWithImageUrl, FormDataType } from '../type'
+
+//  ユーザー確認メッセージを表示するBlockを作成
+export const createMessageBlockForAuthorization = (
+  data: SignUpAuthorizationDataTypeDataWithImageUrl
+) => {
+  const { name, email, password, certificationImageUrl, uid } = data
   return [
     {
       "type": "header",
       "text": {
         "type": "plain_text",
-        "text": name ? `${name}` : "No name",
+        "text": name ? name : "No name",
       },
     },
     {
       "type": "section",
       "text": {
         "type": "plain_text",
-        "text": email ? `${email}` : "No email address",
+        "text": uid ? uid : "No name",
+      },
+    },
+    {
+      "type": "section",
+      "text": {
+        "type": "plain_text",
+        "text": email ?? "No email address",
       },
       "block_id": "text1",
     },
@@ -24,15 +35,20 @@ export const createMessageBlock = (data: SignUpData) => {
       "type": "section",
       "text": {
         "type": "plain_text",
-        "text": password ? `${password}`: "No password",
+        "text": password ?? "No password",
       },
       "block_id": "text2",
     },
-    // {
-    //   "type": "image",
-    //   "image_url": "http://placekitten.com/700/500",
-    //   "alt_text": "Multiple cute kittens"
-    // },
+    {
+      "type": "image",
+      "title": {
+        "type": "plain_text",
+        "text": "医学生および研修医の証明となるもの",
+        "emoji": true,
+      },
+      "image_url": certificationImageUrl ?? "https://pbs.twimg.com/profile_images/625633822235693056/lNGUneLX_400x400.jpg",
+      "alt_text": "cute cat",
+    },
     {
       "type": "actions",
       "block_id": "is_approved_or_not",
@@ -50,3 +66,32 @@ export const createMessageBlock = (data: SignUpData) => {
     },
   ];
 };
+
+export const createMessageBlockForFormReceiver = (formData: FormDataType) => {
+  const { name, email, content } = formData
+  return [
+    {
+      "type": "header",
+      "text": {
+        "type": "plain_text",
+        "text": name ? name : "No name",
+      },
+    },
+    {
+      "type": "section",
+      "text": {
+        "type": "plain_text",
+        "text": email ?? "No email address",
+      },
+      "block_id": "text1",
+    },
+    {
+      "type": "section",
+      "text": {
+        "type": "plain_text",
+        "text": content ?? "No content text",
+      },
+      "block_id": "text2",
+    },
+  ]
+}
