@@ -1,5 +1,12 @@
 module.exports = {
   images: {
-    domains: ['firebasestorage.googleapis.com'],
+    domains: [
+      'firebasestorage.googleapis.com',
+      'res.cloudinary.com',
+      'i.ytimg.com',
+      'youtube.com',
+      'youbu.be',
+    ],
   },
+  swcMinify: true,
 }
