@@ -1,29 +1,37 @@
-export const depList: { name: string; page: string }[] = [
-  { name: '循環器', page: 'Departments/EachDepart' },
-  { name: '消化器内科', page: 'Departments/EachDepart' },
-  { name: '脳神経内科', page: 'Departments/EachDepart' },
-  { name: '腎臓内科', page: 'Departments/EachDepart' },
-  { name: '血液内科', page: 'Departments/EachDepart' },
-  { name: 'アレルギーリウマチ', page: 'Departments/EachDepart' },
-  { name: '皮膚科', page: 'Departments/EachDepart' },
-  { name: '泌尿器科', page: 'Departments/EachDepart' },
-  { name: '脳神経外科', page: 'Departments/EachDepart' },
-  { name: '心臓血管外科', page: 'Departments/EachDepart' },
-  { name: '呼吸器外科', page: 'Departments/EachDepart' },
-  { name: '消化器外科', page: 'Departments/EachDepart' },
-  { name: '移植外科', page: 'Departments/EachDepart' },
-  { name: '緩和ケア', page: 'Departments/EachDepart' },
-  { name: '放射線診断・IVR科', page: 'Departments/Radiology' },
-  { name: '総合診断科', page: 'Departments/EachDepart' },
-  { name: '感染症科', page: 'Departments/EachDepart' },
-  { name: '救急', page: 'Departments/EachDepart' },
-  { name: '整形外科', page: 'Departments/EachDepart' },
-  { name: '耳鼻咽喉科', page: 'Departments/EachDepart' },
-  { name: '麻酔科', page: 'Departments/EachDepart' },
-  { name: '眼科', page: 'Departments/EachDepart' },
-  { name: '形成外科', page: 'Departments/EachDepart' },
-  { name: '公衆衛生', page: 'Departments/EachDepart' },
-  { name: '外傷外科', page: 'Departments/EachDepart' },
+export const departmentCategoryList: string[] = [
+  '消化器内科',
+  '消化器外科',
+  '循環器内科',
+  '心臓血管外科',
+  '整形外科',
+  '小児内科',
+  '小児外科',
+  '脳神経内科',
+  '脳神経外科',
+  '乳腺甲状腺外科',
+  '内分泌代謝内科',
+  '放射線診断科',
+  '放射線腫瘍科',
+  '総合診療科',
+  '総合内科',
+  '眼科',
+  '皮膚科',
+  '耳鼻咽喉科',
+  '産婦人科',
+  '救急科',
+  '腎泌尿器外科',
+  '腎臓内科',
+  '精神科',
+  '感染症科',
+  '膠原病内科',
+  '血液内科',
+  '呼吸器内科',
+  '呼吸器外科',
+  '腫瘍内科',
+  '形成外科',
+  'リハビリテーション科',
+  '病理科',
+  '歯科口腔外科',
 ]
 
 export const sliderData = [
@@ -57,3 +65,159 @@ export const sliderData = [
     videoId: 'OZpv_AcPCKg',
   },
 ]
+
+export const prefectureList = [
+  '北海道',
+  '青森県',
+  '岩手県',
+  '宮城県',
+  '秋田県',
+  '山形県',
+  '福島県',
+  '茨城県',
+  '栃木県',
+  '群馬県',
+  '埼玉県',
+  '千葉県',
+  '東京都',
+  '神奈川県',
+  '新潟県',
+  '富山県',
+  '石川県',
+  '福井県',
+  '山梨県',
+  '長野県',
+  '岐阜県',
+  '静岡県',
+  '愛知県',
+  '三重県',
+  '滋賀県',
+  '京都府',
+  '大阪府',
+  '兵庫県',
+  '奈良県',
+  '和歌山県',
+  '鳥取県',
+  '島根県',
+  '岡山県',
+  '広島県',
+  '山口県',
+  '徳島県',
+  '香川県',
+  '愛媛県',
+  '高知県',
+  '福岡県',
+  '佐賀県',
+  '長崎県',
+  '熊本県',
+  '大分県',
+  '宮崎県',
+  '鹿児島県',
+  '沖縄県',
+]
+export const universityList = [
+  '旭川医科大学',
+  '北海道大学',
+  '弘前大学',
+  '東北大学',
+  '秋田大学',
+  '山形大学',
+  '筑波大学',
+  '群馬大学',
+  '千葉大学',
+  '東京大学',
+  '東京医科歯科大学',
+  '新潟大学',
+  '富山大学',
+  '金沢大学',
+  '福井大学',
+  '山梨大学',
+  '信州大学',
+  '岐阜大学',
+  '浜松医科大学',
+  '名古屋大学',
+  '三重大学',
+  '滋賀医科大学',
+  '京都大学',
+  '大阪大学',
+  '神戸大学',
+  '鳥取大学',
+  '島根大学',
+  '岡山大学',
+  '広島大学',
+  '山口大学',
+  '徳島大学',
+  '香川大学',
+  '愛媛大学',
+  '高知大学',
+  '九州大学',
+  '佐賀大学',
+  '長崎大学',
+  '熊本大学',
+  '大分大学',
+  '宮崎大学',
+  '鹿児島大学',
+  '琉球大学',
+  '札幌医科大学',
+  '福島県立医科大学',
+  '横浜市立大学',
+  '名古屋市立大学',
+  '京都府立医科大学',
+  '大阪市立大学',
+  '奈良県立医科大学',
+  '和歌山県立医科大学',
+  '岩手医科大学',
+  '自治医科大学',
+  '獨協医科大学',
+  '埼玉医科大学',
+  '北里大学',
+  '杏林大学',
+  '慶應義塾大学',
+  '順天堂大学',
+  '昭和大学',
+  '帝京大学',
+  '東海大学',
+  '東京医科大学',
+  '東京慈恵会医科大学',
+  '東京女子医科大学',
+  '東邦大学',
+  '日本大学',
+  '日本医科大学',
+  '聖マリアンナ医科大学',
+  '金沢医科大学',
+  '愛知医科大学',
+  '藤田保健衛生大学',
+  '大阪医科大学',
+  '関西医科大学',
+  '近畿大学',
+  '兵庫医科大学',
+  '川崎医科大学',
+  '久留米大学',
+  '産業医科大学',
+  '福岡大学',
+  '防衛医科大学校',
+]
+
+export const gradeList = ['1', '2', '3', '4', '5', '6']
+
+export const month = [
+  '1',
+  '2',
+  '3',
+  '4',
+  '5',
+  '6',
+  '7',
+  '8',
+  '9',
+  '10',
+  '11',
+  '12',
+]
+
+const newYear = new Date().getFullYear()
+const years: string[] = []
+for (let i = 0; i < 10; i++) {
+  years.push(String(newYear - i))
+}
+export { years }
